Refactor ComplexUserController.java to UserBookController & UserRoleController
