+++ src/main/java/se/iths/librarysystem/controller/UserBookController.java
@@ -3,7 +3,9 @@ package se.iths.librarysystem.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import se.iths.librarysystem.dto.*;
+import se.iths.librarysystem.dto.Book;
+import se.iths.librarysystem.dto.Isbn;
+import se.iths.librarysystem.dto.Task;
 import se.iths.librarysystem.entity.BookEntity;
 import se.iths.librarysystem.entity.TaskEntity;
 import se.iths.librarysystem.entity.UserEntity;
@@ -13,40 +15,29 @@ import se.iths.librarysystem.service.BookService;
 import se.iths.librarysystem.service.TaskService;
 import se.iths.librarysystem.service.UserService;
 import se.iths.librarysystem.validatorservice.BookValidator;
-import se.iths.librarysystem.validatorservice.RoleValidator;
 import se.iths.librarysystem.validatorservice.UserValidator;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("api/users")
-public class ComplexUserController {
+public class UserBookController {
 
     private final UserService userService;
     private final UserValidator userValidator;
     private final BookValidator bookValidator;
-    private final RoleValidator roleValidator;
     private final BookService bookService;
     private final TaskService taskService;
     private final QueueHandler queueHandler;
 
-
-    public ComplexUserController(UserService userService, UserValidator userValidator,
-                                 BookValidator bookValidator, RoleValidator roleValidator, BookService bookService,
-                                 TaskService taskService, QueueHandler queueHandler) {
+    public UserBookController(UserService userService, UserValidator userValidator, BookValidator bookValidator,
+                              BookService bookService, TaskService taskService, QueueHandler queueHandler) {
         this.userService = userService;
         this.userValidator = userValidator;
         this.bookValidator = bookValidator;
-        this.roleValidator = roleValidator;
         this.bookService = bookService;
         this.taskService = taskService;
         this.queueHandler = queueHandler;
-    }
-
-    @GetMapping("{id}/role")
-    public ResponseEntity<List<Role>> getUserRole(@PathVariable Long id) {
-        List<Role> roles = userService.getUserRoles(id);
-        return new ResponseEntity<>(roles, HttpStatus.OK);
     }
 
     @GetMapping("{id}/books")
@@ -61,6 +52,7 @@ public class ComplexUserController {
         BookEntity book = bookService.findById(bookId).orElseThrow(() -> new IdNotFoundException("book", bookId));
         bookValidator.hasUser(book, user);
         book.removeBorrower();
+        bookService.updateBook(book);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
@@ -75,14 +67,6 @@ public class ComplexUserController {
         Task task = queueHandler.sendToQueue(savedTask);
 
         return new ResponseEntity<>(task, HttpStatus.ACCEPTED);
-    }
-
-    @PatchMapping("{userId}/role/{roleId}")
-    public ResponseEntity<UserWithRole> updateUserRole(@PathVariable Long userId, @PathVariable Long roleId) {
-        userValidator.validId(userId);
-        roleValidator.validId(roleId);
-        UserWithRole user = userService.addRoleToUser(userId, roleId);
-        return new ResponseEntity<>(user, HttpStatus.OK);
     }
 
 }
